backend/requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
ecdsa==0.17.0
Markdown==3.3.4
psycopg2-binary==2.8.6
pyasn1==0.4.8
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
rsa==4.7.2
six==1.16.0
sqlparse==0.4.1
